+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 2020-09-14
 
+- [new的简易实现](./docs/JavaScript/simple-new.md)  
 - [call, apply, bind的简易实现](./docs/JavaScript/call-apply-bind.md)  
 
 ### 2020-09-10